plug-controller/requirements.txt
flask==3.0.3
paho-mqtt==2.1.0
Requests==2.32.3